pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = "AutoGen‑powered multi‑agent assistant for large codebases."
authors = ["Your Name <you@example.com>"]
packages = [{include = "agent", from = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python    = "^3.12"
pydantic  = "^2.7"
openai    = "^1.30"
chromadb  = "^0.5"
tiktoken  = "^0.7"
autogen-agentchat = "^0.4"
aiohttp   = "^3.9"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
mypy  = "^1.10"
pytest = "^8.2"

[tool.poetry.scripts]
agent = "agent.cli:main"

[build-system]
requires      = ["poetry-core>=1.8"]
build-backend = "poetry.core.masonry.api"
